requirements.txt
annotated-types==0.7.0
anyio==4.4.0
av==10.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==3.20.0
deepl==1.18.0
distro==1.9.0
exceptiongroup==1.2.1
faster-whisper==0.9.0
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.10.0
future==0.18.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
mpmath==1.3.0
numpy==1.26.1
onnxruntime==1.16.1
packaging==23.2
protobuf==4.25.0
pydantic==2.8.0
pydantic_core==2.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.14.1
tqdm==4.66.1
typing_extensions==4.12.2
urllib3==2.2.2
